+++ ash/src/extensions/khr/maintenance4.rs
@@ -3,6 +3,7 @@ use crate::{Device, Instance};
 use std::ffi::CStr;
 use std::mem;
 
+/// <https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VK_KHR_maintenance4.html>
 #[derive(Clone)]
 pub struct Maintenance4 {
     handle: vk::Device,